reggie-tools/pyproject.toml
[build-system]
requires = [
    "uv_build>=0.8.23,<0.9.0",
]
build-backend = "uv_build"

[project]
name = "reggie-tools"
version = "0.0.1+g63ae48a"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk",
    "databricks-connect<=17.1",
    "requests",
    "urllib3",
    "reggie-core @ file://${PROJECT_ROOT}/../reggie-core",
]

[tool.uv.sources.reggie-core]
workspace = true
